a2a/agentic-workshop/pyproject.toml
[project]
name = "agentic-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncclick>=8.1.8",
    "build>=1.3.0",
    "deprecated>=1.2.18",
    "google-adk[eval]>=1.10.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.108.0",
    "google-cloud-spanner>=3.56.0",
    "google-genai>=1.29.0",
    "humanize>=4.12.3",
    "mcp[cli]>=1.12.4",
    "nest-asyncio>=1.6.0",
    "python-dateutil>=2.9.0.post0",
]
